requirements.txt
Flask
pillow
requests
flask-cors
gunicorn